feat(types): add missing analytics and dashboard types to central module
Closes #349
