+++ src/web/src/types/dashboard.ts
@@ -1,0 +1,33 @@
+/**
+ * Dashboard domain types
+ *
+ * Types for dashboard statistics and overview data.
+ * Aligned with C# DTOs in src/Koinon.Application/DTOs/
+ *
+ * Note: TypeScript uses camelCase for properties while C# uses PascalCase.
+ * JSON serialization handles the casing transformation automatically.
+ * Date fields use ISO 8601 string format for C# DateOnly/DateTime types.
+ */
+
+/**
+ * Upcoming schedule summary for dashboard display
+ */
+export interface UpcomingScheduleDto {
+  idKey: string;
+  name: string;
+  nextOccurrence: string;
+  minutesUntilCheckIn: number;
+}
+
+/**
+ * Dashboard statistics overview
+ */
+export interface DashboardStatsDto {
+  totalPeople: number;
+  totalFamilies: number;
+  activeGroups: number;
+  todayCheckIns: number;
+  lastWeekCheckIns: number;
+  activeSchedules: number;
+  upcomingSchedules: UpcomingScheduleDto[];
+}
