+++ src/web/src/types/index.ts
@@ -10,9 +10,11 @@
  */
 
 // Domain types
+export * from './analytics';
 export * from './authorized-pickup';
 export * from './checkin-extended';
 export * from './communication';
+export * from './dashboard';
 export * from './files';
 export * from './followup';
 export * from './giving';
