+++ tools/graph/backend-graph.json
@@ -1,6 +1,6 @@
 {
   "version": "1.0",
-  "generated_at": "2026-01-02T13:53:56.322311+00:00",
+  "generated_at": "2026-01-02T14:22:57.748419+00:00",
   "entities": {
     "Attendance": {
       "name": "Attendance",
